project_3d/requirements.txt
absl-py==2.1.0
altair==5.3.0
array_record==0.5.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.1.0
blinker==1.8.2
cachetools==5.3.3
cattrs==23.1.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dm-tree==0.1.8
docstring-to-markdown==0.12
docstring_parser==0.16
etils==1.7.0
exceptiongroup==1.1.3
executing==2.0.1
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
imageio==2.34.1
immutabledict==4.2.0
importlib-metadata==6.8.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jedi-language-server==0.41.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
llvmlite==0.43.0
lsprotocol==2023.0.0b1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
microvenv==2023.5.post1
ml-dtypes==0.3.2
namex==0.0.8
nest-asyncio==1.6.0
nibabel==5.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-cupti-cu12==12.3.101
nvidia-cuda-nvcc-cu12==12.3.107
nvidia-cuda-nvrtc-cu12==12.3.107
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==8.9.7.29
nvidia-cufft-cu12==11.0.12.1
nvidia-curand-cu12==10.3.4.107
nvidia-cusolver-cu12==11.5.4.101
nvidia-cusparse-cu12==12.2.0.103
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
opencv-python==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pydeck==0.9.1
pygls==1.0.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
simple_parsing==0.1.5
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.36.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.0rc0
tensorflow-datasets==4.9.6
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-metadata==1.15.0
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typeguard==3.0.2
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.17.0
